hospital section skeletal complete

+++ sql/schema.sql
@@ -16,7 +16,7 @@ CREATE TABLE blood_units (
     unit_id INT PRIMARY KEY AUTO_INCREMENT,
     donor_id INT,
     blood_type ENUM('A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-') NOT NULL,
-    volume DECIMAL(5,2) NOT NULL,
+    volume INT(5) NOT NULL,
     expiration_date DATE NOT NULL,
     status ENUM('available', 'unavailable', 'donated') NOT NULL DEFAULT 'available',
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -38,7 +38,7 @@ CREATE TABLE blood_requests (
     request_id INT PRIMARY KEY AUTO_INCREMENT,
     recipient_id INT,
     blood_type ENUM('A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-') NOT NULL,
-    volume DECIMAL(5,2) NOT NULL,
+    volume INT(5) NOT NULL,
     request_date DATE NOT NULL,
     status ENUM('pending', 'fulfilled', 'cancelled') NOT NULL DEFAULT 'pending',
     FOREIGN KEY (recipient_id) REFERENCES users(user_id) ON DELETE SET NULL
